Splash now gets fiats and saves to storage if necessary

+++ app/src/main/java/com/example/cryptomanager_v2/data/db/DBFiat.kt
@@ -7,5 +7,5 @@ import androidx.room.PrimaryKey
 @Entity
 data class DBFiat(
     @PrimaryKey val name: String,
-    @ColumnInfo(name = "rate") val rate: Int
+    @ColumnInfo(name = "rate") val rate: Double
 )
